Move async helper types into their own module
