+++ src/dolt_annex/commands/filestore/insert.py
@@ -4,7 +4,7 @@
 from plumbum import cli # type: ignore
 
 from dolt_annex.commands import CommandGroup, SubCommand
-from dolt_annex.datatypes.async_utils import maybe_await
+from dolt_annex.datatypes.async_types import maybe_await
 from dolt_annex.datatypes.config import Config
 from dolt_annex.datatypes.repo import Repo
 from dolt_annex.file_keys import get_file_key_type
