+++ src/dolt_annex/filestore/file_handles.py
@@ -13,9 +13,9 @@ import aiofiles
 from fs.base import FS as FileSystem
 
 from dolt_annex.datatypes import FileKey
-from dolt_annex.datatypes.file_io import WritableFileObject, FileInfo, ReadableFileObject
+from dolt_annex.datatypes.async_types import ReadableFileObject, WritableFileObject
+from dolt_annex.datatypes.file_io import FileInfo
 from dolt_annex.filestore.cas import ContentAddressableStorage
-
 
 CHUNK_SIZE = 8092
