+++ src/dolt_annex/filestore/measure.py
@@ -8,12 +8,13 @@ from io import TextIOWrapper
 from pathlib import Path
 from typing_extensions import override
 
+from dolt_annex.datatypes.async_types import AsyncContextManager, AwaitOrEnter, MaybeAwaitable, ReadableFileObject, ReadableStream
 from dolt_annex.datatypes.async_utils import Result
 from dolt_annex.datatypes.config import Config
-from dolt_annex.datatypes.file_io import FileInfo, ReadableStream
+from dolt_annex.datatypes.file_io import FileInfo
 from dolt_annex.file_keys import FileKey
 from dolt_annex.filestore import FileStore
-from dolt_annex.filestore.base import FileStoreModel, MaybeAwaitable, ReadableFileObject
+from dolt_annex.filestore.base import FileStoreModel
 
 @dataclass
 class Measure(FileStore):
