+++ src/dolt_annex/test_util/__init__.py
@@ -14,7 +14,7 @@ from plumbum import cli # type: ignore[import]
 import pytest
 
 from dolt_annex.application import Application
-from dolt_annex.datatypes.async_utils import maybe_await
+from dolt_annex.datatypes.async_types import maybe_await
 from dolt_annex.datatypes.config import Config, DoltConfig, UserConfig
 from dolt_annex.datatypes.repo import RepoModel
 from dolt_annex.datatypes.table import DatasetSchema, FileTableSchema
